Rewrite sources as TypeScript code
