+++ client/src/typescript-module.ts
@@ -1,0 +1,11 @@
+export function getMessageO(): string {
+  return 'MessageO';
+}
+
+export const messageP = 'MessageP';
+
+export default class MessageQ {
+  get(): string {
+    return 'MessageQ';
+  }
+}
